+++ src/components/alert/bl-alert-item.css
@@ -1,0 +1,65 @@
+:host {
+  flex: auto;
+
+  --bl-alert-item-left-content-padding: 0;
+  --bl-alert-item-main-color: var(--bl-color-success);
+  --bl-alert-item-main-bg-color: var(--bl-color-success-background);
+  --bl-alert-item-description-font-size: var(--bl-font-size-m);
+}
+
+bl-button {
+  --bl-button-content-color: var(--bl-alert-item-main-color);
+}
+
+.description {
+  font-size: var(--bl-alert-item-description-font-size);
+  line-height: var(--bl-alert-item-description-font-size);
+}
+
+.content {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex: auto;
+}
+
+.left-content {
+  display: flex;
+  align-items: center;
+  padding: var(--bl-alert-item-left-content-padding)
+}
+
+.icon {
+  margin-right: var(--bl-size-2xs);
+  color: var(--bl-alert-item-main-color);
+}
+
+.text-content {
+  display: flex;
+  flex-direction: column;
+}
+
+.title {
+  color:var(--bl-alert-item-main-color);
+  font: var(--bl-font-title-3-medium);
+  margin-bottom: var(--bl-size-3xs);
+}
+
+:host([title]){
+  --bl-alert-item-description-font-size: var(--bl-font-size-s);
+}
+
+:host([variant='info']){
+  --bl-alert-item-main-color: var(--bl-color-primary);
+  --bl-alert-item-main-bg-color: var(--bl-color-accent-primary-background);
+}
+
+:host([variant='warning']){
+  --bl-alert-item-main-color: var(--bl-color-warning);
+  --bl-alert-item-main-bg-color: var(--bl-color-warning-background);
+}
+
+:host([variant='error']){
+  --bl-alert-item-main-color: var(--bl-color-danger);
+  --bl-alert-item-main-bg-color: var(--bl-color-danger-background);
+}
