--- src/components/alert/bl-alert.css
@@ -1,117 +1,0 @@
-:host {
-  --bl-alert-padding: var(--bl-size-m);
-  --bl-alert-main-color: var(--bl-color-success);
-  --bl-alert-main-bg-color: var(--bl-color-success-background);
-  --bl-alert-description-font-size: var(--bl-font-size-m);
-}
-
-.alert {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background-color: var(--bl-alert-main-bg-color);
-  color: var(--bl-color-content-primary);
-  border: 1px solid var(--bl-alert-main-color);
-  border-radius: var(--bl-size-3xs);
-  padding: 0 var(--bl-alert-padding);
-}
-
-.alert bl-button {
-  --bl-button-content-color: var(--bl-alert-main-color);
-}
-
-.description {
-  font-size: var(--bl-alert-description-font-size);
-  line-height: var(--bl-alert-description-font-size);
-}
-
-.content {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex: auto;
-}
-
-.left-content {
-  display: flex;
-  align-items: center;
-  padding: calc(var(--bl-alert-padding) - var(--bl-size-4xs) / 2 ) 0;
-}
-
-.icon {
-  margin-right: var(--bl-size-2xs);
-  color: var(--bl-alert-main-color);
-}
-
-.close {
-  display: flex;
-  padding: var(--bl-size-2xs);
-  cursor: pointer;
-}
-
-.text-content {
-  display: flex;
-  flex-direction: column;
-}
-
-.title {
-  color:var(--bl-alert-main-color);
-  font: var(--bl-font-title-3-medium);
-  margin-bottom: var(--bl-size-3xs);
-}
-
-.actions {
-  display: flex;
-  align-items: center;
-}
-
-.pagination {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: var(--bl-size-2xs);
-  user-select: none;
-}
-
-.counter {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: var(--bl-size-3xs);
-  margin-left: var(--bl-size-3xs);
-  font: var(--bl-font-title-3-medium);
-}
-
-.counter span:not(:first-child) {
-  color: var(--bl-color-content-secondary)
-}
-
-.page-number{
-  min-width: 20px;
-  text-align: center;
-}
-
-.arrow {
-  font: var(--bl-size-xs);
-  color: var(--bl-color-content-primary);
-  cursor: pointer;
-}
-
-:host([title]){
-  --bl-alert-description-font-size: var(--bl-font-size-s);
-}
-
-:host([variant='info']){
-  --bl-alert-main-color: var(--bl-color-primary);
-  --bl-alert-main-bg-color: var(--bl-color-accent-primary-background);
-}
-
-:host([variant='warning']){
-  --bl-alert-main-color: var(--bl-color-warning);
-  --bl-alert-main-bg-color: var(--bl-color-warning-background);
-}
-
-:host([variant='error']){
-  --bl-alert-main-color: var(--bl-color-danger);
-  --bl-alert-main-bg-color: var(--bl-color-danger-background);
-}
